+++ src/clock_init.rs
@@ -269,7 +269,7 @@ pub fn probe_clocks() -> SystemClocks {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, not(target_arch = "riscv32")))]
 mod tests {
     use super::*;